+++ packages/solv/CHANGELOG.md
@@ -1,4 +1,8 @@
 # @gabrielhicks/solv
+## 5.6.8
+### Monitoring Module
+Use local files
+
 ## 5.6.7
 ### Monitoring Module
 Use local files